Refactor navigation and footer components; replace Navbar with new structure and styles, and implement Footer component for consistent layout.
