+++ client/src/App.scss
@@ -16,16 +16,3 @@
     color: $secondary-color-black; // Ensure h1 inside header also uses black for contrast
   }
 }
-
-footer {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  margin-top: 40px;
-  padding: 20px;
-  background-color: $secondary-color-gray;
-  color: $secondary-color-black;
-  font-size: 0.9rem;
-  z-index: 1000;
-}
