+++ client/src/components/Navigation/footer.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import "../../styles/footer.scss";
+
+function Footer() {
+  return (
+    <footer>
+      <p>&copy; {new Date().getFullYear()} The Old South Bake House</p>
+    </footer>
+  );
+}
+
+export default Footer;
