+++ client/src/styles/navBar.scss
@@ -1,4 +1,4 @@
-@import "../styles/main.scss";
+@import "./main.scss";
 
 .navbar {
   display: flex;
